Fix column size adjustment inconsistencies, use ResizeObserver API
The CSS for the table needs to be managed in JS because the table and
tile components use virtual scrolling, which only renders rows and tiles
that are visible. This normally leads to the columns resizing as you
scroll, so this works around it by watching for th width changes from an
opacified thead, then once the DOM computes it, applying it to the fixed
thead clone and table rows.